fixed the error

+++ contracts/poseidon/src/interface.rs
@@ -6,7 +6,7 @@ mod callable {
 
     sol_interface! {
         interface PoseidonInterface {
-            function hash(uint[2] memory inputs) external view returns (uint hash);
+            function hash(uint256[2] memory inputs) external view returns (uint256 hash);
         }
     }
 }
